requirements.txt
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
groq==0.4.1
mistralai==0.0.12
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
tqdm==4.65.0
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0